+++ src/globals.py
@@ -1,2 +1,2 @@
-build=83
-type='stable'
+build=84
+type='dev'